Create interface to add abbreviates
The plugin now can add abbreviates to a given file, using a specific
command for it. The command AutoAbbrevAddAbbrev gets the lhs and rhs of
the abbrev file, and adds this abbrev to the abbrev file, to the current
session (by running the command manually) and replaces all the previous
occurances of lhs into rhs.

There are many options that can be controlled by globals:
* loaded_auto_abbrev - to load or not load the plugin.
* auto_abbrev_use_file - To save the abbreviates to a given file, or
  not.
* auto_abbrev_file_path - The path to the abbreviates file.
* auto_abbrev_should_fix_abbrev - Whether to change the abbrev words, or
  leave them as they are.

Currently, there aren't any documentation on the plugin's actions, it
would be added in a later commit.