Add add_current_word as a new abbrev function
Created a new command: AutoAbbrevAddCurrentWord. This command would add
the word under the cursor as a new abbreviate. This would probably be
the most useful function in the plugin, since it is the most easily
mapped function, that would be easy to use as the change command.

The function gets the word under the cursor, and then replaces it with
a new word, one got interactively from the user.