Add option to source the abbrev file automatically
Added a flag and a simple command, that would allow the user to source
the abbreviation file automatically from inside the plugin, instead of
sourcing it implicitly in some other part of the initialization process.

Now, the abbreviate file can be self contained in the plugin, there
shouldn't be any reason to touch it directly.