Add support for deleting duplicated abbreviates
Added new functions, that would delete abbreviates that already appears
in the abbrev file.

Every time the plugin is loaded, all the existing abbreviates are read
from the file, in that process, every duplicated abbreviate is deleted,
saving only the most recent abbreviate.

Later, each time a new abbreviate is added to the file, the previous
abbreviates are updated, removing the already existing abbrev in case it
already in the dictionary. The dictionary of existed abbreviates are
always updated with the abbrev file, making sure to delete the line
from the file once the abbrev is being removed.

There is a flag that can be toggled to use or not use this option, since
this option is kind of invasive (it deletes abbreviates from the file
automatically), and it might make the plugin slower.