Add reload support to the plugin
Added a function that would support reloading of the abbreviates of the
plugin. This option would be useful in case someone would edit the file
manually, or add abbreviates from more than one instance of vim.

If more than one instance of vim would make changes to the file, doing
actions without reloading it might result in corrupting the file - it
would remove the wrong abbrev in case of duplication, and might write
bad data in case of parallel using of the file. The reload function
would add interface that would let the user use the plugin in multiple
instances of vim, but it still wouldn't keep him safe from corrupting
the file, in case he won't use it.