Add seperate commands for lhs and rhs abbreviates
Added another interface with the user. This interface can be used to add
words that are easy to fix, but hard to type. To use this interface, the
user would have to start by standing on a word with an error that he
wants to replace. He needs to save this word, fix it to the wanted word,
and then save the new word as the abbreviate.

This interface is quite hard to use by commands, but for now I haven't
added any mapping to the plugin. Once the mapping would be added, there
would be mapping for those commands as well, and they would be easier to
use.