Add support for abbreviates deletion
Now, the user would be able to remove abbreviate from the abbreviate
file using API from the plugin.

The command is AutoAbbrevDelAbbrev and it would remove the abbreviate
line with the given word from the abbreviate file.

Now, the user won't have any reason to edit the abbreviate files by
himself, all the functions for modifying this file are inside the
plugin.