Add mapping for common functions
Added mapping to the function from the plugin. The functions that got
mappings are the functions that it is logical to call them from within
the file, without more arguments.
Those functions are:
* add_current_word
* add_rhs_word
* add_lhs_word
* reload

Those mappings have normal mode option and visual mode option. The
normal mode option let those functions run on the current word (the word
under the cursor). The visual mode mapping would run the command on the
visual selected word. Those mapping should not work for visual mode
that is not regular visual mode (line visual or block visual).