Fix error regarding the mapping doc in README
There was an error in the README file. All the mappings actually starts with a leader, but the markdown format removed this.
This was probably done due to the fact that markdown regard the string of <leader> as tag, and removed it from the text.

The fix was simply to add escaping before the first '<', and it made the markdown leave the word there.